start match dirs
