+++ src/terms/match/index.js
@@ -1,9 +1,8 @@
 'use strict';
-//
 const syntax = require('./lib/syntax');
 const startHere = require('./lib/startHere');
 const Text = require('../../result/index');
-// const diff = require('./diff');
+const match = require('./lib/index');
 
 const matchMethods = (Terms) => {
 
@@ -15,27 +14,7 @@ const matchMethods = (Terms) => {
       if (!reg) {
         return new Text([], this.lexicon, this.parent);
       }
-      //parse for backwards-compatibility
-      if (typeof reg === 'string') {
-        reg = syntax(reg);
-      }
-      if (verbose) {
-        console.log(reg);
-      }
-      let matches = [];
-      for (let t = 0; t < this.terms.length; t++) {
-        //don't loop through if '^'
-        if (reg[0] && reg[0].starting && t > 0) {
-          break;
-        }
-        let m = startHere(this, t, reg, verbose);
-        if (m) {
-          matches.push(m);
-          //ok, don't try to match these again.
-          let skip = m.length - 1;
-          t += skip; //this could use some work
-        }
-      }
+      let matches = match(this, reg, verbose);
       matches = matches.map((a) => {
         return new Terms(a, this.lexicon, this.refText, this.refTerms);
       });
